refactor(): some more restructuring
